changed initlog to simulate 3 button pushes

+++ background.py
@@ -235,7 +235,7 @@ def avgcomp():  # positions are in complications
         for row in reader:
             if row[2] != 0:
                 zd1.append(row[2])
-    a = int(sum(zd1) / len(zd1))
+    a = sum(zd1) / len(zd1)
     complications(2)
     compnumber(2, a)
     comptext(2, c2t)
